% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_casesVcontrols.R
\name{casesVcontrols}
\alias{casesVcontrols}
\title{This function performs the casesVcontrols exon analysis on 2 AllInfo.gz files.}
\usage{
casesVcontrols(
  caseList,
  controlList,
  chrom_file,
  numThreads = 10,
  annotation_path,
  ci_low = 0.05,
  ci_high = 0.95,
  min_reads = 10,
  zipping_function = "zcat",
  OL_fraction = 0.8,
  CellTypeFile,
  OutputDir = "OutputDir"
)
}
\arguments{
\item{caseList}{file with path to "case" AllInfo.gz}

\item{controlList}{file with path to "control" AllInfo.gz}

\item{chrom_file}{file with list of chromosomes desired for testing}

\item{numThreads}{number of threads to be used}

\item{annotation_path}{path to gencode formatted annotation}

\item{ci_low}{min percent spliced inclusion considered; recommended starting value = 0.05}

\item{min_reads}{minimum number of reads for sum of 2 allInfos for a given exon}

\item{zipping_function}{command for unzipping files; Linux = "zcat"}

\item{OL_fraction}{the fraction of the reads for a given position must be either inclusion or exclusion; recommended starting value = 0.8}

\item{CellTypeFile}{file with list of all celltypes to be considered}

\item{OutputDir}{name of OutputDir. Default is OutputDir}

\item{ci_upper}{max percent spliced inclusion considered; recommended starting value = 0.95}
}
\value{
cases_vs_controls.counts.passed-chi-sq-crit.tab.gz : exon_gene  includedcases excludedcases includedcontrols excludedcontrols

cases_vs_controls.counts.pVal.FDR.LOR.tab.gz

all.altExons.matrix.tab.gz, all.altExons.tab.gz, all.internalExons.tab.gz, CASES.sampleBoth.inc.exc.tab.gz, CONTROLS.sampleBoth.inc.exc.tab.gz

LargeDPSI_exons above |DPSI| > 0.5 , MediumDPSI_exons above |DPSI| > 0.25, sampleBoth.inc.exc.tab. per cell types
}
\description{
Exons must pass chi sqaure crit. to be tested. Those exons are outputted into cases_vs_controls.counts.passed-chi-sq-crit.tab.gz with counts returned.
Exons are also outputted into cases_vs_controls.counts.pVal.FDR.LOR.tab.gz
}
