% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_CreateExampleATACobj.R
\name{Create_Example_ATACobj}
\alias{Create_Example_ATACobj}
\title{create an example seurat object includes the chromatin assay with the fragment files and barcode metrics, condition and cell type per cell will be assigned.
the object wil be names as 'combined' and can be applied as input for the random subsampling}
\usage{
Create_Example_ATACobj(
  example.data.path,
  outDir = "OutputDir",
  harmony = FALSE
)
}
\arguments{
\item{outDir}{the directory for the output file}

\item{harmony}{will perform RunHarmony if set to be TRUE, default = FALSE.}
}
\value{
example atac object
}
\description{
.
}
