% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_DAPeaksByCelltype.R
\name{DAPeaks_ByCelltype}
\alias{DAPeaks_ByCelltype}
\title{Calling differential accessible peaks by comparing two conditions of one specific cell type ####################}
\usage{
DAPeaks_ByCelltype(
  ATACobj,
  annotation.gr = NULL,
  AssayName = "ATAC",
  condition.query,
  celltypeA,
  celltypeB,
  cellnum = 500,
  peaknum = 5000,
  MinCellRatio = 0.02,
  random.repeats = 10,
  harmony = FALSE,
  outputDir,
  savePeakRobj = FALSE
)
}
\arguments{
\item{ATACobj}{the object with chromatin assay created.}

\item{annotation.gr}{A set of GRanges containing annotations for the genome used, the default setting is NULL. If supplied, it will be used as input for CreateChromatinAssay.}

\item{AssayName}{The assay names of the chromatin assay, default = "ATAC"}

\item{condition.query}{the condition in query.}

\item{celltypeA}{the cell type A for comparison.}

\item{celltypeB}{the cell type B for comparison.}

\item{cellnum}{the number of cells to be subsampled, default = 500.}

\item{peaknum}{the number of peaks to be subsampled, default = 5000.}

\item{MinCellRatio}{Only test peaks that are detected in a minimum fraction of MinCellRatio cells in either of the two conditions, the default setting is 0.02.}

\item{random.repeats}{Random subsampling times to be performed, default = 10.}

\item{harmony}{will perform RunHarmony if set to be TRUE, default = FALSE.}

\item{outputDir}{Directory path to the output filee, which will be created if not existed.}

\item{savePeakRobj}{The peaks called by MACS2 for each subsampling will be stored as assay ‘peaks’. It will be saved as Robj for downstream analysis, default = FALSE.}
}
\value{
downsampled ATAC cells
}
\description{
Calling differential accessible peaks by comparing two conditions of one specific cell type ####################
}
