% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_DownloadReferenceFiles.R
\name{DownloadRefs}
\alias{DownloadRefs}
\title{Exons are also outputted into cases_vs_controls.counts.pVal.FDR.LOR.tab.gz}
\usage{
DownloadRefs()
}
\value{
downloaded example files
}
\description{
Exons are also outputted into cases_vs_controls.counts.pVal.FDR.LOR.tab.gz
}
