% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_downsampleExons.R
\name{downsampleExonsAndIterate}
\alias{downsampleExonsAndIterate}
\title{Downsamples number exons and then repeats to generate a distribution of significant events}
\usage{
downsampleExonsAndIterate(Num_Exons_Selected, Num_Repeats, example = FALSE)
}
\arguments{
\item{Num_Exons_Selected}{Number of exons to select to calculate \% significant}

\item{Num_Repeats}{Number of times to loop over randomly selecting exons}
}
\value{
Percent_Sig_Downsampled.tab
}
\description{
.
}
