% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_downsampleReads.R
\name{downsampleReads}
\alias{downsampleReads}
\title{Downsamples number of reads per exon in CasesVControls output: cases_vs_controls.counts.passed-chi-sq-crit.tab.gz}
\usage{
downsampleReads(Num_Downsampled_Reads = 10, example = FALSE)
}
\arguments{
\item{Num_Downsampled_Reads}{Num of minimum reads per exon and how many reads will be downsampled.}
}
\value{
Sampling_DPSI_Table.tab

Correlation plot
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
