% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_ViolinPlot.R
\name{ViolinPlot}
\alias{ViolinPlot}
\title{Downsamples number of reads, exons, and then repeats to generate a distribution of significant events from CasesVControls output: cases_vs_controls.counts.passed-chi-sq-crit.tab.gz}
\usage{
ViolinPlot()
}
\value{
violin plot from output of downsampleExonsAndIterate.R
}
\description{
This function converts input temperatures in Fahrenheit to Celsius.
}
